renamed module dictIO.collections to dictIO.dict
